+++ mcp_bridge/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from endpoints import router as endpointRouter
 from mcp_endpoints import router as api_router
-from database.db_endpoints import router as db_router
+from web.router import router as web_router
 from health import router as healthRouter
 from lifespan import lifespan
 from openapi_tags import tags_metadata
@@ -16,7 +16,7 @@ app = FastAPI(
 app.include_router(endpointRouter)
 app.include_router(api_router)
 app.include_router(healthRouter)
-app.include_router(db_router)
+app.include_router(web_router)
 
 if __name__ == "__main__":
     import uvicorn
