+++ mcp_bridge/web/templates/base.jinja
