+++ mcp_bridge/web/templates/index.jinja
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "base.jinja" %}
 
 {% block content %}
 <div class="row">
